pyproject.toml
[tool.poetry]
name = "simplechat"
version = "0.1.0"
description = ""
authors = ["Daniele"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
starlette = "^0.41.3"
uvicorn = "^0.32.1"
black = "^24.10.0"
toolhouse = "^1.3.0"
openai = "^1.57.0"
anthropic = "^0.40.0"
python-dotenv = "^1.0.1"
h2 = "^4.1.0"
uvloop = "^0.21.0"
httptools = "^0.6.4"
hypercorn = "^0.17.3"
httpx = "^0.28.1"
certifi = "^2024.8.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
